refactor(calendar): extract calendar_watch_relay crate and drop relay serving from email_service
The relay's serving side moves to its own dev-only service, so
email_service reverts to the plain webhook handler; the wire model, SSE
parsing, secret comparison, and subscriber config move to a shared crate
consumed by both the tunnel service and the subscriber worker.
