+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "crates/bots",
   "crates/broadcast",
   "crates/calendar_events",
+  "crates/calendar_watch_relay",
   "crates/channel_bots",
   "crates/client/cache-core",
   "crates/client/cache-idb",
@@ -60,6 +61,7 @@ members = [
   "crates/workspace-hack",
   "services/ai_projections_refresh_handler",
   "services/authentication_service",
+  "services/calendar_event_local_tunnel",
   "services/call_recording_preview_handler",
   "services/connection_gateway",
   "services/contacts_service",
