+++ crates/calendar_watch_relay/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+edition = "2024"
+name = "calendar_watch_relay"
+publish = false
+version = "0.1.0"
+
+[dependencies]
+macro_env_var = { path = "../macro_env_var" }
+serde = { workspace = true, features = ["derive"] }
+sha2 = { workspace = true }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
+
+[dev-dependencies]
+serde_json = { workspace = true }
