+++ services/email_service/Cargo.toml
@@ -77,6 +77,7 @@ sfs_delete = []
 
 [dependencies]
 calendar_events = { path = "../../crates/calendar_events", features = ["google", "inbound", "postgres"] }
+calendar_watch_relay = { path = "../../crates/calendar_watch_relay" }
 ammonia = { workspace = true }
 anyhow = { workspace = true }
 authentication_service_client = { path = "../../crates/authentication_service_client" }
