+++ services/email_service/src/openapi.rs
@@ -5,7 +5,6 @@ mod backfill_completion_service;
 mod backfill_init_service;
 mod calendar_outbox;
 mod calendar_tokens;
-mod calendar_watch_relay;
 mod config;
 mod outbound;
 mod pubsub;
